Mostly coding style corrections.
One change:
 * Send and Cancel, renamed property to match case with WD (sendId -> sendid)

Style corrections based on:
 * Checkstyle
 * Findbugs
 * Javadoc
 * Unneccessary casts


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/sandbox/scxml/trunk@327091 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/scxml/SCXMLHelper.java
@@ -81,7 +81,7 @@ public final class SCXMLHelper {
      */
     public static Set getAncestorClosure(final Set states,
             final Set upperBounds) {
-        Set closure = new HashSet((int) (states.size() * 2));
+        Set closure = new HashSet(states.size() * 2);
         for (Iterator i = states.iterator(); i.hasNext();) {
             TransitionTarget tt = (TransitionTarget) i.next();
             closure.add(tt);
